Extract Handlebars::Transform to its own file
