+++ lib/ruby-handlebars.rb
@@ -2,6 +2,7 @@ require_relative "ruby-handlebars/version"
 require_relative "ruby-handlebars/context"
 require_relative "ruby-handlebars/parser"
 require_relative "ruby-handlebars/tree"
+require_relative "ruby-handlebars/transform"
 require_relative "ruby-handlebars/template"
 require_relative "ruby-handlebars/helper"
 require_relative "ruby-handlebars/helpers/register_default_helpers"
