+++ lib/ruby-handlebars/tree.rb
@@ -108,88 +108,4 @@ module Handlebars
 
     end
   end
-
-  class Transform < Parslet::Transform
-    rule(template_content: simple(:content)) {Tree::TemplateContent.new(content)}
-    rule(replaced_unsafe_item: simple(:item)) {Tree::EscapedReplacement.new(item)}
-    rule(replaced_safe_item: simple(:item)) {Tree::Replacement.new(item)}
-    rule(str_content: simple(:content)) {Tree::String.new(content)}
-    rule(parameter_name: simple(:name)) {Tree::Parameter.new(name)}
-
-    rule(
-      unsafe_helper_name: simple(:name),
-      parameters: subtree(:parameters)
-    ) {
-      Tree::EscapedHelper.new(name, parameters)
-    }
-
-    rule(
-      safe_helper_name: simple(:name),
-      parameters: subtree(:parameters)
-    ) {
-      Tree::Helper.new(name, parameters)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      block_items: subtree(:block_items),
-    ) {
-      Tree::Helper.new(name, [], block_items)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      block_items: subtree(:block_items),
-      else_block_items: subtree(:else_block_items)
-    ) {
-      Tree::Helper.new(name, [], block_items, else_block_items)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      parameters: subtree(:parameters),
-      block_items: subtree(:block_items),
-    ) {
-      Tree::Helper.new(name, parameters, block_items)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      parameters: subtree(:parameters),
-      block_items: subtree(:block_items),
-      else_block_items: subtree(:else_block_items)
-    ) {
-      Tree::Helper.new(name, parameters, block_items, else_block_items)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      parameters: subtree(:parameters),
-      as_parameters: subtree(:as_parameters),
-      block_items: subtree(:block_items),
-    ) {
-      Tree::AsHelper.new(name, parameters, as_parameters, block_items)
-    }
-
-    rule(
-      helper_name: simple(:name),
-      parameters: subtree(:parameters),
-      as_parameters: subtree(:as_parameters),
-      block_items: subtree(:block_items),
-      else_block_items: subtree(:else_block_items)
-    ) {
-      Tree::AsHelper.new(name, parameters, as_parameters, block_items, else_block_items)
-    }
-    
-    rule(
-      partial_name: simple(:partial_name),
-      arguments: subtree(:arguments)
-    ) {
-      Tree::PartialWithArgs.new(partial_name, arguments)
-    }
-
-    rule(partial_name: simple(:partial_name)) {Tree::Partial.new(partial_name)}
-    rule(block_items: subtree(:block_items)) {Tree::Block.new(block_items)}
-    rule(else_block_items: subtree(:else_block_items)) {Tree::Block.new(block_items)}
-  end
 end
